@@ -9,7 +9,7 @@ maintainers = [
 ]
 description = "Monty is the missing complement to Python."
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 classifiers = [
     "Programming Language :: Python :: 3",
     "Development Status :: 4 - Beta",
@@ -20,9 +20,9 @@ classifiers = [
 ]
 dependencies = [
   "ruamel.yaml",
-  "numpy<2.0.0",
+  "numpy",
 ]
-version = "2024.10.21"
+version = "2024.12.10"
 
 [project.optional-dependencies]
 ci = [
@@ -31,6 +31,7 @@ ci = [
   "pytest>=8",
   "pytest-cov>=4",
   "types-requests",
+  "pymongo"
 ]
 # dev is for "dev" module, not for development
 dev = ["ipython"]
@@ -39,12 +40,14 @@ docs = [
     "sphinx_rtd_theme",
 ]
 json = [
-  "bson",
+  "pymongo",
   "orjson>=3.6.1",
   "pandas",
   "pydantic",
-  "pint",
-  "torch",
+  # https://github.com/hgrecco/pint/issues/2065
+  "pint; python_version<'3.13'",
+  # Note: need torch>=2.3.0 for numpy 2 # 719
+  "torch; python_version<'3.13'",  # python 3.13 not supported yet
 ]
 multiprocessing = ["tqdm"]
 optional = ["monty[dev,json,multiprocessing,serialization]"]
@@ -57,10 +60,9 @@ include = ["monty", "monty.*"]
 
 [tool.black]
 line-length = 120
-target-version = ['py39']
+target-version = ["py310"]
 include = '\.pyi?$'
 exclude = '''
-
 (
   /(
       \.eggs         # exclude a few common directories in the
@@ -85,8 +87,6 @@ branch = true
 exclude_also = [
     "@deprecated",
     "def __repr__",
-    "if 0:",
-    "if __name__ == .__main__.:",
     "if self.debug:",
     "if settings.DEBUG",
     "pragma: no cover",
@@ -94,17 +94,17 @@ exclude_also = [
     "raise NotImplementedError",
     "show_plot",
     "if TYPE_CHECKING:",
-    "if typing.TYPE_CHECKING:",
     "except ImportError:"
 ]
 
 [tool.mypy]
 ignore_missing_imports = true
 
-[tool.ruff]
-lint.select = [
-  "I",  #isort
+[tool.ruff.lint]
+select = [
+  "I",  # isort
 ]
 
-lint.isort.required-imports = ["from __future__ import annotations"]
-lint.isort.known-first-party = ["monty"]
+[tool.ruff.lint.isort]
+required-imports = ["from __future__ import annotations"]
+known-first-party = ["monty"]
